django backend/requirements.txt
django==6.0.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
firebase-admin==7.2.0
psycopg2-binary==2.9.11
django-cors-headers==4.9.0

